Updated Readme - FAQ section.
Hi, I have added this section from one of your closed and SOLVED issues. I was searching for this already for two times (I always forget what to do since I am not updating Go version that often) and having this directly in the readme could be very helpful for everybody.